rename package dnascissors to geneditid; rename webapp to geneditidapp; rename projects folder GEPIDs to PROJECTS
