+++ .gitignore
@@ -32,6 +32,6 @@ __pycache__
 geneditid.sqlite
 
 # Projects folder
-GEPIDs/
+PROJECTS/
 
 # Miscellaneous
