+++ python/geneditidapp/MANIFEST.in
@@ -1,0 +1,2 @@
+include *.txt *.ini *.cfg *.rst
+recursive-include geneditidapp *.ico *.png *.css *.gif *.jpg *.pt *.txt *.mak *.mako *.js *.html *.xml *.jinja2
