+++ python/geneditidapp/development.ini
@@ -36,7 +36,7 @@ listen = localhost:8080
 ###
 
 [loggers]
-keys = root, dnascissors, geneditid, sqlalchemy, webapp
+keys = root, geneditid, geneditidapp, sqlalchemy
 
 [handlers]
 keys = console, filelog
@@ -48,20 +48,15 @@ keys = generic
 level = DEBUG
 handlers = console, filelog
 
-[logger_webapp]
-level = DEBUG
-handlers = console, filelog
-qualname = webapp
-
-[logger_dnascissors]
-level = DEBUG
-handlers = console, filelog
-qualname = dnascissors
-
 [logger_geneditid]
 level = DEBUG
 handlers = console, filelog
 qualname = geneditid
+
+[logger_geneditidapp]
+level = DEBUG
+handlers = console, filelog
+qualname = geneditidapp
 
 [logger_sqlalchemy]
 level = WARN
@@ -79,7 +74,7 @@ formatter = generic
 
 [handler_filelog]
 class = handlers.RotatingFileHandler
-args = ('%(here)s/../../GEPIDs/geneditid_webapp.log','a')
+args = ('%(here)s/../../PROJECTS/geneditidapp.log','a')
 level = INFO
 formatter = generic
