+++ python/geneditidapp/production.ini
@@ -28,7 +28,7 @@ listen = *:6543
 ###
 
 [loggers]
-keys = root, dnascissors, geneditid, sqlalchemy, wsgi
+keys = root, geneditid, geneditidapp, sqlalchemy, wsgi
 
 [handlers]
 keys = console, filelog, accesslog
@@ -40,15 +40,15 @@ keys = generic
 level = WARN
 handlers = console, filelog
 
-[logger_dnascissors]
-level = DEBUG
-handlers = console, filelog
-qualname = dnascissors
-
 [logger_geneditid]
 level = DEBUG
 handlers = console, filelog
 qualname = geneditid
+
+[logger_geneditidapp]
+level = DEBUG
+handlers = console, filelog
+qualname = geneditidapp
 
 [logger_sqlalchemy]
 level = WARN
@@ -72,7 +72,7 @@ formatter = generic
 
 [handler_filelog]
 class = logging.handlers.RotatingFileHandler
-args = ('%(here)s/webapp.log','a')
+args = ('%(here)s/geneditidapp.log','a')
 level = INFO
 formatter = generic
