+++ python/geneditidapp/pyramid.wsgi
@@ -1,4 +1,4 @@
 from pyramid.paster import get_app, setup_logging
-ini_path = '/home/ge/GenEditID/python/webapp/production.ini'
+ini_path = '/home/ge/GenEditID/python/geneditidapp/production.ini'
 setup_logging(ini_path)
 application = get_app(ini_path, 'main')
