+++ python/geneditidapp/pytest.ini
@@ -1,3 +1,3 @@
 [pytest]
-testpaths = webapp
+testpaths = geneditidapp
 python_files = *.py
