+++ python/geneditidapp/static/dnascissors.css
