+++ python/geneditidapp/static/pyramid-16x16.png
