+++ python/geneditidapp/static/pyramid.png
