+++ python/geneditidapp/static/sequence.js
