+++ python/geneditidapp/templates/layout.pt
@@ -9,7 +9,7 @@
     <link rel="stylesheet" type="text/css" href="https://cdn.datatables.net/1.10.15/css/dataTables.bootstrap.min.css"/>
     <link rel="stylesheet" type="text/css" href="https://cdn.datatables.net/buttons/1.3.1/css/buttons.bootstrap.min.css"/>
 
-    <link href='/static/dnascissors.css' rel='stylesheet' type='text/css' />
+    <link href='/static/geneditid.css' rel='stylesheet' type='text/css' />
 
     <!-- For Plotly
     ======================================================================== -->
