+++ python/geneditidapp/views/home.py
@@ -8,8 +8,8 @@ from pyramid.httpexceptions import HTTPFound
 
 from sqlalchemy.exc import DBAPIError
 
-from dnascissors.config import cfg
-from dnascissors.model import Project
+from geneditid.config import cfg
+from geneditid.model import Project
 
 
 class HomeViews(object):
@@ -126,7 +126,7 @@ class HomeViews(object):
     Pyramid is having a problem using your SQL database.  The problem
     might be caused by one of the following things:
 
-    1.  You may need to run the "initialize_webapp_db" script
+    1.  You may need to run the "initialize_geneditidapp_db" script
         to initialize your database tables.  Check your virtual
         environment's "bin" directory for this script and try to run it.
