+++ python/scripts/create_db.py
@@ -3,8 +3,8 @@ Python3 file of the genome-editing project
 Created by Anne Pajon @pajanne on 08/03/2017
 """
 import os
-from dnascissors.model import Base
-from dnascissors.config import cfg
+from geneditid.model import Base
+from geneditid.config import cfg
 from sqlalchemy import create_engine
 
 engine = create_engine(cfg['DATABASE_URI'])
