+++ python/scripts/load_cell_growth.py
@@ -1,8 +1,8 @@
 import sqlalchemy
 
-from dnascissors.config import cfg
-from dnascissors.model import Base
-from dnascissors.loader import CellGrowthLoader
+from geneditid.config import cfg
+from geneditid.model import Base
+from geneditid.loader import CellGrowthLoader
 
 import log as logger
