+++ python/scripts/load_layout.py
@@ -1,8 +1,8 @@
 import sqlalchemy
 
-from dnascissors.config import cfg
-from dnascissors.model import Base
-from dnascissors.loader import LayoutLoader
+from geneditid.config import cfg
+from geneditid.model import Base
+from geneditid.loader import LayoutLoader
 
 import log as logger
