+++ python/scripts/load_protein_abundance.py
@@ -1,8 +1,8 @@
 import sqlalchemy
 
-from dnascissors.config import cfg
-from dnascissors.model import Base
-from dnascissors.loader import ProteinAbundanceLoader
+from geneditid.config import cfg
+from geneditid.model import Base
+from geneditid.loader import ProteinAbundanceLoader
 
 import log as logger
