+++ python/scripts/load_ref_data.py
@@ -1,9 +1,9 @@
 import os
 import sqlalchemy
 
-from dnascissors.config import cfg
-from dnascissors.model import Base
-from dnascissors.loader import RefLoader
+from geneditid.config import cfg
+from geneditid.model import Base
+from geneditid.loader import RefLoader
 
 import log as logger
