--- python/webapp/.coveragerc
@@ -1,3 +1,0 @@
-[run]
-source = webapp
-omit = webapp/test*
