--- python/webapp/MANIFEST.in
@@ -1,2 +1,0 @@
-include *.txt *.ini *.cfg *.rst
-recursive-include webapp *.ico *.png *.css *.gif *.jpg *.pt *.txt *.mak *.mako *.js *.html *.xml *.jinja2
